+++ src/main/java/nextstep/auth/authentication/AuthenticationInterceptor.java
@@ -1,7 +1,7 @@
 package nextstep.auth.authentication;
 
+import nextstep.auth.authentication.converter.AuthenticationConverter;
 import nextstep.auth.context.Authentication;
-import nextstep.member.domain.LoginMember;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
