+++ src/main/java/nextstep/auth/authentication/converter/SessionAuthenticationConverter.java
@@ -1,4 +1,6 @@
-package nextstep.auth.authentication;
+package nextstep.auth.authentication.converter;
+
+import nextstep.auth.authentication.AuthenticationToken;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
