+++ src/test/java/nextstep/subway/unit/auth/SessionAuthenticationInterceptorTest.java
@@ -1,31 +1,29 @@
-package nextstep.subway.unit;
+package nextstep.subway.unit.auth;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.authentication.UserDetailsService;
 import nextstep.auth.authentication.*;
+import nextstep.auth.authentication.converter.AuthenticationConverter;
+import nextstep.auth.authentication.converter.SessionAuthenticationConverter;
 import nextstep.auth.context.Authentication;
-import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenResponse;
+import nextstep.auth.context.SecurityContext;
 import nextstep.member.application.CustomUserDetailsService;
 import nextstep.member.domain.LoginMember;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
-import static nextstep.subway.unit.AuthFixture.*;
+import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
+import static nextstep.subway.unit.auth.AuthFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
-class TokenAuthenticationInterceptorTest {
+class SessionAuthenticationInterceptorTest {
     UserDetailsService userDetailsService;
-    JwtTokenProvider jwtTokenProvider;
     AuthenticationConverter converter;
     AuthenticationInterceptor interceptor;
     MockHttpServletRequest request;
@@ -34,11 +32,9 @@ class TokenAuthenticationInterceptorTest {
     @BeforeEach
     void setUp() throws IOException {
         userDetailsService = mock(CustomUserDetailsService.class);
-        jwtTokenProvider = mock(JwtTokenProvider.class);
-        converter = new TokenAuthenticationConverter();
-        interceptor = new TokenAuthenticationInterceptor(
-                userDetailsService, converter, jwtTokenProvider, new ObjectMapper());
-        request = createTokenMockRequest();
+        converter = new SessionAuthenticationConverter();
+        interceptor = new SessionAuthenticationInterceptor(userDetailsService, converter);
+        request = createSessionMockRequest();
         response = createMockResponse();
     }
 
@@ -60,15 +56,16 @@ class TokenAuthenticationInterceptorTest {
     void afterAuthentication() throws IOException {
         // given
         Authentication authentication = createAuthentication();
-        given(jwtTokenProvider.createToken(anyString())).willReturn(JWT_TOKEN);
 
         // when
         interceptor.afterAuthentication(request, response, authentication);
 
         // then
+        HttpSession session = request.getSession();
+        SecurityContext securityContext = (SecurityContext) session.getAttribute(SPRING_SECURITY_CONTEXT_KEY);
+
         assertThat(response.getStatus()).isEqualTo(HttpStatus.OK.value());
-        assertThat(response.getContentType()).isEqualTo(MediaType.APPLICATION_JSON_VALUE);
-        assertThat(response.getContentAsString()).isEqualTo(new ObjectMapper().writeValueAsString(new TokenResponse(JWT_TOKEN)));
+        assertThat(securityContext.getAuthentication()).isEqualTo(authentication);
     }
 
     @Test
@@ -76,14 +73,17 @@ class TokenAuthenticationInterceptorTest {
         // given
         given(userDetailsService.loadUserByUsername(EMAIL))
                 .willReturn(new LoginMember(1L, EMAIL, PASSWORD, 20));
-        given(jwtTokenProvider.createToken(anyString())).willReturn(JWT_TOKEN);
 
         // when
         interceptor.preHandle(request, response, new Object());
 
         // then
+        HttpSession session = request.getSession();
+        SecurityContext securityContext = (SecurityContext) session.getAttribute(SPRING_SECURITY_CONTEXT_KEY);
+        LoginMember loginMember = (LoginMember) securityContext.getAuthentication().getPrincipal();
+
         assertThat(response.getStatus()).isEqualTo(HttpStatus.OK.value());
-        assertThat(response.getContentType()).isEqualTo(MediaType.APPLICATION_JSON_VALUE);
-        assertThat(response.getContentAsString()).isEqualTo(new ObjectMapper().writeValueAsString(new TokenResponse(JWT_TOKEN)));
+        assertThat(loginMember.getEmail()).isEqualTo(EMAIL);
+        assertThat(loginMember.getPassword()).isEqualTo(PASSWORD);
     }
 }
