+++ src/test/java/nextstep/subway/unit/auth/TokenAuthenticationConverterTest.java
@@ -1,23 +1,23 @@
-package nextstep.subway.unit;
+package nextstep.subway.unit.auth;
 
-import nextstep.auth.authentication.AuthenticationConverter;
+import nextstep.auth.authentication.converter.AuthenticationConverter;
 import nextstep.auth.authentication.AuthenticationToken;
-import nextstep.auth.authentication.SessionAuthenticationConverter;
+import nextstep.auth.authentication.converter.TokenAuthenticationConverter;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 
 import java.io.IOException;
 
-import static nextstep.subway.unit.AuthFixture.*;
+import static nextstep.subway.unit.auth.AuthFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class SessionAuthenticationConverterTest {
-    AuthenticationConverter converter = new SessionAuthenticationConverter();
+class TokenAuthenticationConverterTest {
+    AuthenticationConverter converter = new TokenAuthenticationConverter();
 
     @Test
     void convert() throws IOException {
         // given
-        MockHttpServletRequest request = createSessionMockRequest();
+        MockHttpServletRequest request = createTokenMockRequest();
 
         // when
         AuthenticationToken token = converter.convert(request);
